Just more comments
Going to take a look and see if I can eliminate some things later to make things faster. Specifically if we get rid of the decompressFile function and have that logic just in main it will likely speed it up a decent amount. I know it won't look as pretty but I just wanna go fast!